requirements.txt
langgraph>=1.0.7
langgraph-checkpoint-sqlite>=3.0.0
langchain-openai>=1.1.7
python-dotenv>=1.0.0
httpx>=0.27.0
chainlit>=1.0.0
pillow>=10.0.0
aiosqlite
SQLAlchemy
